services/requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
numpy>=1.24.0
scipy>=1.10.0
torch>=2.0.0
torchaudio>=2.0.0
openvoice>=1.0.0
melo-tts>=1.0.0

